# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: my_customTools API
  description: |
    媒体转码工具集后端 API 文档。

    包含视频转码、图片处理、动图转换、文档转换等功能。

    ## 认证方式
    使用 JWT Bearer Token 认证，在请求头中添加：
    ```
    Authorization: Bearer <token>
    ```

    ## 响应格式
    所有接口统一返回格式：
    ```json
    {
      "code": 0,
      "msg": "ok",
      "data": {}
    }
    ```
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api
    description: 开发环境
  - url: http://production-server/api
    description: 生产环境
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Token 认证
  schemas:
...
    ApiResponse:
      type: object
      required:
        - code
        - msg
        - data
      properties:
        code:
          type: integer
          description: 业务状态码，0 表示成功
          example: 0
        msg:
          type: string
          description: 响应消息
          example: "ok"
        data:
          type: object
          description: 响应数据
...
    CreateUploadSessionRequest:
      type: object
      required:
        - fileName
        - fileSize
        - type
      properties:
        fileName:
          type: string
        fileSize:
          type: integer
          minimum: 1
        type:
          type: string
          enum: [video, img, anim]
    UploadProgress:
      type: object
      properties:
        uploadId:
          type: string
        uploadedChunks:
          type: integer
        totalChunks:
          type: integer
        progress:
          type: integer
          minimum: 0
          maximum: 100

    # ========== 队列相关 ==========
...
    SystemConfig:
      type: object
      properties:
        video:
          type: object
          properties:
            parallelLimit:
              type: integer
            maxFileSize:
              type: integer
            allowedInputFormats:
              type: array
              items:
                type: string
        img:
          type: object
          properties:
            parallelLimit:
              type: integer
            maxFileSize:
              type: integer
            allowedInputFormats:
              type: array
              items:
                type: string
        upload:
          type: object
          properties:
            chunkSize:
              type: integer
            maxParallelUploads:
              type: integer
        storage:
          type: object
          properties:
            type:
              type: string
              enum: [local, s3]
            uploadDir:
              type: string
            outputDir:
              type: string
            autoClean:
              type: boolean
            cleanDays:
              type: integer
        updatedAt:
          type: string
          format: date-time

    # ========== 菜单相关 ==========
...
paths:
...
  /upload/session:
    post:
      tags: [Upload]
      summary: 创建上传会话
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUploadSessionRequest'
      responses:
        '200':
          description: 会话创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          uploadId:
                            type: string
                          totalChunks:
                            type: integer
                          chunkSize:
                            type: integer
        '400':
          description: 参数错误
  /upload/chunk:
    post:
      tags: [Upload]
      summary: 上传分片
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - uploadId
                - chunkIndex
                - file
              properties:
                uploadId:
                  type: string
                chunkIndex:
                  type: integer
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: 上传成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UploadProgress'
        '400':
          description: 参数错误或上传失败
  /upload/complete:
    post:
      tags: [Upload]
      summary: 完成上传
      description: 合并所有分片，返回文件路径
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - uploadId
              properties:
                uploadId:
                  type: string
      responses:
        '200':
          description: 上传完成
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          filePath:
                            type: string
                          fileName:
                            type: string
        '400':
          description: 参数错误或合并失败
  /upload/progress/{uploadId}:
    get:
      tags: [Upload]
      summary: 获取上传进度
      parameters:
        - name: uploadId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UploadProgress'
        '404':
          description: 上传会话不存在
  /upload/cancel/{uploadId}:
    delete:
      tags: [Upload]
      summary: 取消上传
      parameters:
        - name: uploadId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 取消成功
        '400':
          description: 取消失败
...
  /config:
    get:
      tags: [Config]
      summary: 获取系统配置
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SystemConfig'

    put:
      tags: [Config]
      summary: 更新系统配置
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: 部分配置更新
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SystemConfig'
        '400':
          description: 参数错误
...